chezmoi: updated scripts

+++ run_onchange_install_apt_tools.sh
@@ -4,5 +4,11 @@ sudo apt update
 sudo apt install bat eza fd-find git-delta ripgrep hyperfine -y
 sudo apt autoremove -y
 
+if [ ! -f ~/.local/bin/fd ]; then
 ln -s $(which fdfind) ~/.local/bin/fd
+fi
+
+if [ ! -f ~/.local/bin/bat ]; then
 ln -s $(which batcat) ~/.local/bin/bat
+fi
+
